@@ -1,15 +1,20 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
   <PropertyGroup>
-    <TargetFramework>netcoreapp2.2</TargetFramework>
+    <TargetFramework>netcoreapp3.1</TargetFramework>
     <OutputType>Exe</OutputType>
     <Version>1.13.13</Version>
     <AssemblyName>MarginTrading.TradingHistory</AssemblyName>
     <RootNamespace>MarginTrading.TradingHistory</RootNamespace>
-    <LangVersion>7.2</LangVersion>
+    <LangVersion>latest</LangVersion>
+    <AspNetCoreHostingModel>OutOfProcess</AspNetCoreHostingModel>
+    <AspNetCoreModuleName>AspNetCoreModuleV2</AspNetCoreModuleName>
   </PropertyGroup>
   <PropertyGroup>
-    <DocumentationFile>bin\Release\netcoreapp2.0\MarginTrading.TradingHistory.xml</DocumentationFile>
+    <DocumentationFile>bin\Release\netcoreapp3.1\MarginTrading.TradingHistory.xml</DocumentationFile>
   </PropertyGroup>
+  <ItemGroup>
+    <FrameworkReference Include="Microsoft.AspNetCore.App" />
+  </ItemGroup>
   <ItemGroup>
     <None Update="Dockerfile">
       <CopyToPublishDirectory>PreserveNewest</CopyToPublishDirectory>
