@@ -2,16 +2,14 @@
 // See the LICENSE file in the project root for more information.
 
 using System;
-using System.IO;
-using System.Net;
-using System.Security.Cryptography.X509Certificates;
 using System.Threading;
 using System.Threading.Tasks;
+using Autofac.Extensions.DependencyInjection;
 using JetBrains.Annotations;
 using MarginTrading.TradingHistory.Core.Services;
-using Microsoft.AspNetCore;
 using Microsoft.AspNetCore.Hosting;
 using Microsoft.Extensions.Configuration;
+using Microsoft.Extensions.Hosting;
 using Microsoft.Extensions.PlatformAbstractions;
 
 namespace MarginTrading.TradingHistory
@@ -19,7 +17,7 @@ namespace MarginTrading.TradingHistory
     [UsedImplicitly]
     internal sealed class Program
     {
-        internal static IWebHost Host { get; private set; }
+        internal static IHost AppHost { get; private set; }
         public static string EnvInfo => Environment.GetEnvironmentVariable("ENV_INFO");
         
         public static async Task Main(string[] args)
@@ -45,13 +43,20 @@ namespace MarginTrading.TradingHistory
                         .AddEnvironmentVariables()
                         .Build();
 
-                    Host = WebHost.CreateDefaultBuilder()
-                        .UseConfiguration(configuration)
-                        .UseStartup<Startup>()
-                        .UseApplicationInsights()
+                    AppHost = Host.CreateDefaultBuilder(args)
+                        .UseServiceProviderFactory(new AutofacServiceProviderFactory())
+                        .ConfigureWebHostDefaults(webBuilder =>
+                        {
+                            webBuilder.ConfigureKestrel(serverOptions =>
+                                {
+                                    // Set properties and call methods on options
+                                })
+                                .UseConfiguration(configuration)
+                                .UseStartup<Startup>();
+                        })
                         .Build();
-
-                    await Host.RunAsync();
+                    
+                    await AppHost.RunAsync();
                 }
                 catch (Exception e)
                 {
